add graph transformer method

+++ gnn_model_train.py
@@ -15,6 +15,7 @@ import warnings
 
 from torch_geometric.utils import dense_to_sparse
 
+from model.GraphConv import load_GraphConv_model, GraphConv_model
 from model.GraphTransformer import GraphTransformer_model, load_GraphTransforer_model
 
 warnings.filterwarnings("ignore")
@@ -130,7 +131,20 @@ if __name__ == "__main__":
         else:
             gnn_model, pre_output = GraphTransformer_model(data, nhid, dropout, lr, weight_decay, gcn_layer, heads_num,
                                                            epoch, device)
-            file_path = os.path.join(model_save_path, 'GraphTransformer.pth')
+            file_path = os.path.join(model_save_path, 'graphTransformer_model.pth')
+            torch.save(gnn_model.state_dict(), file_path)
+            gnn_model.eval()
+    elif not gnn_model and test_model == 'GraphConv':
+        file_path = os.path.join(model_save_path, 'graphConv_model.pth')
+        if os.path.exists(file_path):
+            gnn_model = load_GraphConv_model(file_path, data, nhid, dropout, device, lr, weight_decay, gcn_layer)
+            dense_adj = torch.tensor(adj.toarray())
+            norm_adj = normalize_adj(dense_adj)
+            edge_index, edge_weight = dense_to_sparse(norm_adj)
+            pre_output = gnn_model.forward(torch.tensor(features.toarray()), edge_index, edge_weight=edge_weight)
+        else:
+            gnn_model, pre_output = GraphConv_model(data, nhid, dropout, lr, weight_decay, gcn_layer, epoch, device)
+            file_path = os.path.join(model_save_path, 'graphConv_model.pth')
             torch.save(gnn_model.state_dict(), file_path)
             gnn_model.eval()
